requirements.txt
mcp>=0.9.0
httpx[http2]>=0.25.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
uvicorn>=0.27.0
starlette>=0.36.0
sse-starlette>=1.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

